parser/requirements.txt
black == 21.6b0 
colorama == 0.4.4
docker == 5.0.3
elasticsearch == 8.2.2
pefile == 2022.5.30
requests == 2.25.1
watchdog == 2.1.6
pycryptodomex == 3.12.0
py7zr == 0.17.2
kafka-python == 2.0.2
redpanda == 0.5.1
timeloop == 1.0.2
service == 0.6.0
pyelftools == 0.28
oletools == 0.60.1 
flare-floss == 2.0.0
yara-python == 4.2.0
python_magic == 0.4.27